// src/app/shared/turnstile/turnstile.interfaces.ts
export interface TurnstileOptions {
    sitekey: string;
    action?: string;
    cData?: string;
    callback?: (token: string) => void;
    'error-callback'?: (errorCode: string) => boolean;
    'expired-callback'?: () => void;
    theme?: 'light' | 'dark' | 'auto';
    tabindex?: number;
    appearance?: 'always' | 'execute' | 'interaction-only';
  }